:card_file_box: Update data type

+++ Week 08/CaseStudyPt4/create_price_table_and_sales_table.sql
@@ -1,28 +1,28 @@
 CREATE TABLE prices
 (
-    id                     TINYINT UNSIGNED NOT NULL AUTO_INCREMENT,
-    just_java_endless      FLOAT   UNSIGNED NOT NULL,
-    cafe_au_lait_single    FLOAT   UNSIGNED NOT NULL,
-    cafe_au_lait_double    FLOAT   UNSIGNED NOT NULL,
-    iced_cappuccino_single FLOAT   UNSIGNED NOT NULL,
-    iced_cappuccino_double FLOAT   UNSIGNED NOT NULL,
+    id                     TINYINT      UNSIGNED NOT NULL AUTO_INCREMENT,
+    just_java_endless      DECIMAL(9,2) UNSIGNED NOT NULL,
+    cafe_au_lait_single    DECIMAL(9,2) UNSIGNED NOT NULL,
+    cafe_au_lait_double    DECIMAL(9,2) UNSIGNED NOT NULL,
+    iced_cappuccino_single DECIMAL(9,2) UNSIGNED NOT NULL,
+    iced_cappuccino_double DECIMAL(9,2) UNSIGNED NOT NULL,
 
     PRIMARY KEY (id)
 );
 
 CREATE TABLE sales
 (
-    id                       INT      UNSIGNED           NOT NULL AUTO_INCREMENT,
-    just_java_qty            SMALLINT UNSIGNED DEFAULT 0 NOT NULL,
-    just_java_price          FLOAT    UNSIGNED DEFAULT 0 NOT NULL,
-    just_java_subtotal       FLOAT    UNSIGNED DEFAULT 0 NOT NULL,
-    cafe_au_lait_qty         SMALLINT UNSIGNED DEFAULT 0 NOT NULL,
-    cafe_au_lait_price       FLOAT    UNSIGNED DEFAULT 0 NOT NULL,
-    cafe_au_lait_subtotal    FLOAT    UNSIGNED DEFAULT 0 NOT NULL,
-    iced_cappuccino_qty      SMALLINT UNSIGNED DEFAULT 0 NOT NULL,
-    iced_cappuccino_price    FLOAT    UNSIGNED DEFAULT 0 NOT NULL,
-    iced_cappuccino_subtotal FLOAT    UNSIGNED DEFAULT 0 NOT NULL,
-    total                    FLOAT    UNSIGNED DEFAULT 0 NOT NULL,
+    id                       INT          UNSIGNED           NOT NULL AUTO_INCREMENT,
+    just_java_qty            SMALLINT     UNSIGNED DEFAULT 0 NOT NULL,
+    just_java_price          DECIMAL(9,2) UNSIGNED DEFAULT 0 NOT NULL,
+    just_java_subtotal       DECIMAL(9,2) UNSIGNED DEFAULT 0 NOT NULL,
+    cafe_au_lait_qty         SMALLINT     UNSIGNED DEFAULT 0 NOT NULL,
+    cafe_au_lait_price       DECIMAL(9,2) UNSIGNED DEFAULT 0 NOT NULL,
+    cafe_au_lait_subtotal    DECIMAL(9,2) UNSIGNED DEFAULT 0 NOT NULL,
+    iced_cappuccino_qty      SMALLINT     UNSIGNED DEFAULT 0 NOT NULL,
+    iced_cappuccino_price    DECIMAL(9,2) UNSIGNED DEFAULT 0 NOT NULL,
+    iced_cappuccino_subtotal DECIMAL(9,2) UNSIGNED DEFAULT 0 NOT NULL,
+    total                    DECIMAL(9,2) UNSIGNED DEFAULT 0 NOT NULL,
 
     PRIMARY KEY (id)
 );
